backend_ml/requirements.txt
flask
flask-cors
pandas
numpy
scikit-learn
joblib
python-dotenv
supabase
